:recycle: extract util functions to a separate file
