+++ lib/equality_assertion_strategy.js
@@ -1,30 +1,6 @@
 'use strict';
 
-const assert = require('assert');
-
-function isCyclic(obj) {
-  if (obj === null) { return false; }
-  const seenObjects = [];
-  
-  function detect(obj) {
-    if (typeof obj === 'object') {
-      if (seenObjects.includes(obj)) return true;
-      seenObjects.push(obj);
-      return !!Object.keys(obj).find(key => detect(obj[key]));
-    }
-    return false;
-  }
-  
-  return detect(obj);
-}
-
-function notNullOrUndefined(object) {
-  return object !== undefined && object !== null;
-}
-
-function hasMethodDefined(object, method) {
-  return notNullOrUndefined(object) && typeof object[method] === 'function';
-}
+const Utils = require('./utils');
 
 const EqualityAssertionStrategy = {
   availableStrategies() {
@@ -91,7 +67,7 @@ const CustomPropertyName = {
   },
   
   _comparisonCanBeMade(actual, expected, criteria) {
-    return hasMethodDefined(actual, criteria) && hasMethodDefined(expected, criteria);
+    return Utils.hasMethodDefined(actual, criteria) && Utils.hasMethodDefined(expected, criteria);
   },
   
   _compareUsingCustomCriteria(actual, expected, criteria) {
@@ -113,7 +89,7 @@ const ObjectWithEqualsProperty = {
   __proto__: EqualityAssertionStrategy,
   
   appliesTo(actual, expected) {
-    return hasMethodDefined(actual, 'equals')
+    return Utils.hasMethodDefined(actual, 'equals')
   },
   
   evaluate(actual, expected) {
@@ -128,7 +104,7 @@ const ObjectWithCyclicReference = {
   __proto__: EqualityAssertionStrategy,
   
   appliesTo(actual, expected) {
-    return isCyclic(actual) || isCyclic(expected);
+    return Utils.isCyclic(actual) || Utils.isCyclic(expected);
   },
   
   evaluate(actual, expected) {
@@ -147,14 +123,8 @@ const DefaultEquality = {
   },
   
   evaluate(actual, expected) {
-    let comparisonResult = true;
-    try {
-      assert.deepStrictEqual(actual, expected);
-    } catch (_assertionError) {
-      comparisonResult = false;
-    }
     return {
-      comparisonResult: comparisonResult,
+      comparisonResult: Utils.deepStrictEqual(actual, expected),
       additionalFailureMessage: ''
     };
   },
